[JENKINS-36501] Move Descriptors as top level items
