+++ src/main/java/org/jenkinsci/plugins/runselector/filters/AndRunFilter.java
@@ -27,6 +27,7 @@ package org.jenkinsci.plugins.runselector.filters;
 import hudson.Extension;
 import hudson.model.Run;
 import org.jenkinsci.plugins.runselector.RunFilter;
+import org.jenkinsci.plugins.runselector.RunFilterDescriptor;
 import org.jenkinsci.plugins.runselector.context.RunSelectorPickContext;
 import org.kohsuke.stapler.DataBoundConstructor;
