+++ src/main/java/org/jenkinsci/plugins/runselector/selectors/TriggeringRunSelector.java
@@ -32,6 +32,7 @@ import hudson.model.Job;
 import hudson.model.Run;
 import net.sf.json.JSONObject;
 import org.jenkinsci.plugins.runselector.RunSelector;
+import org.jenkinsci.plugins.runselector.RunSelectorDescriptor;
 import org.jenkinsci.plugins.runselector.context.RunSelectorPickContext;
 import org.jvnet.localizer.Localizable;
 import org.kohsuke.stapler.DataBoundConstructor;
